+++ RushOrders/docs/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* All display text is now translatable.
 * Improved integration with Generic Mod Config Menu (GMCM):
   * Updated for GMCM 1.5.0.
   * Enabled config UI from the in-game options menu after the save is loaded.